+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 4.7
 - Set select, text, multiple, textarea columns type to text in Disciplineform.php, Referral.fnc.php
 - Remove GPA / MP List program in GPAMPList.php, Menu.php & rosariosis.sql
 - Add CLASS_RANK_CALCULATE_MPS to CONFIG table in Update.fnc.php & rosariosis.sql
+- Add ClassRankMaybeCalculate(), ClassRankCalculateAJAX() & ClassRankCalculateAddMP() in ClassRank.inc.php
 
 Changes in 4.6.2
 ----------------